cleanup & minor changes
mostly working but search bar is still a bit finnicky. the original is like this though so I dunno, I'll try and fix it if I can. the most standout issue is that the search bar doesn't focus when you press s, you have to make an additional keypress and I don't know why. it also doesn't unfocus when you close the search bar.